Move specified and computed angles to submodules
